Finalize refactoring of import_gtfs
