--- gtfspy/__init__.py
@@ -1,1 +1,0 @@
-import gtfspy.util as util
