+++ gtfspy/import_loaders/agency_loader.py
@@ -2,8 +2,7 @@ import os
 import time
 from datetime import datetime
 
-from gtfspy.import_gtfs import decode_six
-from gtfspy.import_loaders.table_loader import TableLoader
+from gtfspy.import_loaders.table_loader import TableLoader, decode_six
 
 
 class AgencyLoader(TableLoader):
