+++ gtfspy/import_loaders/feed_info_loader.py
@@ -1,6 +1,5 @@
 from gtfspy.gtfs import GTFS
-from gtfspy.import_gtfs import decode_six
-from gtfspy.import_loaders.table_loader import TableLoader
+from gtfspy.import_loaders.table_loader import TableLoader, decode_six
 
 
 class FeedInfoLoader(TableLoader):
