+++ gtfspy/import_loaders/frequencies_loader.py
@@ -1,8 +1,7 @@
 import pandas
 
 from gtfspy import util
-from gtfspy.import_gtfs import decode_six
-from gtfspy.import_loaders.table_loader import TableLoader
+from gtfspy.import_loaders.table_loader import TableLoader, decode_six
 
 
 class FrequenciesLoader(TableLoader):
